\name{getTemplate}
\alias{getTemplate}
\title{
Get stan model template code for egg count data analysis
}
\description{
Obtain a template stan model code that can be used in the \strong{eggCounts} package.
}
\usage{
getTemplate()
}
\value{
A character string containing the stan model template.
}
\author{
Craig Wang \email{craig.wang@uzh.ch}
}

\seealso{
\code{\link[eggCounts]{fecr_stanExtra}} for using the model specification.
}
