\docType{package}
\name{eggCountsExtra-package}
\alias{eggCountsExtra}
\alias{eggCountsExtra-package}
\title{Add-Ons for eggCounts Package}
\description{
This package provides additional models and functions for analyzing faecal egg count data to assess anthelmintic efficacy.
}
\details{
\tabular{ll}{
Package: \tab eggCountsExtra\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2018-04-16\cr
License: \tab GPL (>= 3) \cr
LazyLoad: \tab yes\cr
}}

\author{
Craig Wang \email{craig.wang@uzh.ch} \cr
}
\keyword{package}

